refactor: PathController, Service 분리
