+++ src/test/java/subway/line/presentation/LineControllerTest.java
@@ -10,7 +10,6 @@ import org.springframework.boot.test.web.server.LocalServerPort;
 import org.springframework.http.HttpStatus;
 import org.springframework.test.context.jdbc.Sql;
 import subway.line.dto.LineCreateDto;
-import subway.line.dto.ShortestPathRequest;
 import subway.line.service.LineService;
 import subway.section.dto.SectionCreateDto;
 import subway.station.dto.StationCreateDto;
@@ -109,28 +108,5 @@ class LineControllerTest {
           .delete("/lines/" + lineId +"/stations/" + 잠실역_id)
         .then()
           .statusCode(HttpStatus.NO_CONTENT.value());
-  }
-
-  @Test
-  void getShortestPath() {
-    final Long lineId1 = lineService.createLine(new LineCreateDto("1호선"));
-    final Long lineId2 = lineService.createLine(new LineCreateDto("2호선"));
-    final Long lineId3 = lineService.createLine(new LineCreateDto("3호선"));
-    final Long 잠실새내역_id = stationService.create(new StationCreateDto("잠실새내역"));
-    final Long 잠실역_id = stationService.create(new StationCreateDto("잠실역"));
-    final Long 잠실나루역_id = stationService.create(new StationCreateDto("잠실나루역"));
-    lineService.addSection(new SectionCreateDto(lineId1, 잠실새내역_id, 잠실나루역_id, 5));
-    lineService.addSection(new SectionCreateDto(lineId2, 잠실새내역_id, 잠실역_id, 4));
-    lineService.addSection(new SectionCreateDto(lineId3, 잠실나루역_id, 잠실역_id, 100));
-    final ShortestPathRequest shortestPathRequest = new ShortestPathRequest(잠실새내역_id, 잠실역_id);
-
-    RestAssured
-        .given()
-          .body(shortestPathRequest)
-          .contentType(ContentType.JSON)
-        .when()
-          .get("/lines/shortest-path")
-        .then()
-          .statusCode(HttpStatus.OK.value());
   }
 }
